feat: introduce minimal plugin system to offset non-essenial logic
